refactor: delete Backend, BackendImpl, BackendConfig — fully unified Client-Server
Deleted 3 files (-993 lines):
- agent/backend.go (154 lines) — AgentBackend interface moved to lifecycle.go
- agent/backend_impl.go (752 lines) — Backend implementation replaced by Client
- agent/backend_config.go (87 lines) — no longer needed

Updated:
- serverapp/callbacks.go: backend *agent.Backend → ag *agent.Agent
- serverapp/server.go: registerChannels uses ag directly
- serverapp/server_core.go: removed Backend field, uses ag directly
- agent/lifecycle.go: hosts AgentBackend interface + AgentRunner + types

Net: -842 lines in this commit.
